# Clean up obsolete code in BlockBasedTable::PrefetchIndexAndFilterBlocks