--- a/unreleased_history/bug_fixes/cleanup_obsolete_code_in_prefetch_index.md
+++ b/unreleased_history/bug_fixes/cleanup_obsolete_code_in_prefetch_index.md
@@ -1,0 +1,1 @@
+Clean up obsolete fix in `BlockBasedTable::PrefetchIndexAndFilterBlocks` isolated to early 7.0.x releases.