@jaykorean 2025-01-08
PR generally looks good to me. Could you add a release note for this change? I think it's important for users to know in case someone upgrades from a very old version of RocksDB < 7.0.x to the version having this change.

https://github.com/facebook/rocksdb/blob/main/unreleased_history/README.txt should have the instruction for the release note.